chore: remove Sentry in dev + build fixes (lucide, calendar, tailwind)
