+++ app/global-error.tsx
@@ -1,14 +1,13 @@
 "use client";
 
-import * as Sentry from "@sentry/nextjs";
 import NextError from "next/error";
 import { useEffect } from "react";
 
-export default function GlobalError({ error }: { error: Error & { digest?: string } }) {
-  useEffect(() => {
-    Sentry.captureException(error);
-  }, [error]);
-
+export default function GlobalError({
+  error,
+}: {
+  error: Error & { digest?: string };
+}) {
   return (
     <html>
       <body>
