+++ package.json
@@ -24,6 +24,7 @@
     "@radix-ui/react-alert-dialog": "^1.0.4",
     "@radix-ui/react-avatar": "^1.1.1",
     "@radix-ui/react-checkbox": "^1.0.4",
+    "@radix-ui/react-collapsible": "1.1.2",
     "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-dropdown-menu": "^2.0.5",
     "@radix-ui/react-icons": "^1.3.0",
@@ -39,7 +40,6 @@
     "@radix-ui/react-tooltip": "^1.0.6",
     "@react-google-maps/api": "^2.19.3",
     "@react-pdf/renderer": "^3.1.12",
-    "@sentry/nextjs": "^8.25.0",
     "@tanstack/react-query": "^4.29.25",
     "@types/node": "18.16.3",
     "@types/react": "18.2.5",
@@ -47,6 +47,7 @@
     "algoliasearch": "^4.17.1",
     "axios": "^1.4.0",
     "bcrypt": "^5.1.0",
+    "bcryptjs": "2.4.3",
     "canvas": "^2.11.2",
     "class-variance-authority": "^0.6.1",
     "clsx": "^1.2.1",
@@ -57,11 +58,11 @@
     "framer-motion": "^10.12.8",
     "jsbarcode": "^3.11.5",
     "jsprintmanager": "^6.0.0",
-    "lucide-react": "^0.260.0",
+    "lucide-react": "0.544.0",
     "md5": "^2.3.0",
     "next": "^14.2.3",
     "next-auth": "^4.22.1",
-    "next-cloudinary": "^4.15.0",
+    "next-cloudinary": "6.16.0",
     "next-images": "^1.8.5",
     "next-intl": "^2.14.6",
     "nextjs-toploader": "^1.3.2",
@@ -70,6 +71,7 @@
     "qrcode.react": "^3.1.0",
     "react": "^18.3.1",
     "react-datepicker": "^4.14.0",
+    "react-day-picker": "8.10.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.2",
     "react-hot-toast": "^2.4.1",
@@ -97,6 +99,7 @@
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.0",
+    "@types/bcryptjs": "2.4.6",
     "@types/md5": "^2.3.2",
     "@types/pdfjs-dist": "^2.10.378",
     "@types/react-datepicker": "^4.11.2",
