+++ pages/_error.jsx
@@ -1,4 +1,4 @@
-import * as Sentry from "@sentry/nextjs";
+
 import Error from "next/error";
 
 const CustomErrorComponent = (props) => {
